resolved build issues with theming interface + dropdown interface
