+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@gdcorte/react-core-components",
-  "version": "0.7.1",
+  "version": "0.7.4",
   "repository": "https://github.com/Gdcorte/react-core-components.git",
   "homepage": "https://github.com/Gdcorte/react-core-components#readme",
   "bugs": {
