+++ src/components/Dropdown/Hybrid/SubDropdown.tsx
@@ -8,13 +8,12 @@ import {
 import { OutsideClickHandler } from '../../../hooks/OutsideClick';
 import { Carrets } from '../../../icons';
 import { renderBodyDropdown } from '../helper';
-import SubDropdown from './SubDropdown';
 
 interface ClickDropdownMenuProps extends DropdownMenuProps {
     parentRef?: RefObject<HTMLObjectElement>
 }
 
-export const ClickDropdown: FunctionComponent<ClickDropdownMenuProps> = ({
+export const SubDropdown: FunctionComponent<ClickDropdownMenuProps> = ({
     label,
     options,
     listOrientation,
@@ -71,9 +70,9 @@ export const ClickDropdown: FunctionComponent<ClickDropdownMenuProps> = ({
     )
 }
 
-ClickDropdown.defaultProps = {
+SubDropdown.defaultProps = {
     listOrientation: "left",
     showCarret: true,
 }
 
-export default ClickDropdown
+export default SubDropdown
