+++ src/components/Dropdown/index.tsx
@@ -1,3 +1,4 @@
 export {default as ClickDropdown} from './Click'
 export {default as HoverDropdown} from './Hover'
 export {default as HybridDrowdown} from './Hybrid'
+export type { DropdownOptionProps } from './interface'
