+++ src/components/index.tsx
@@ -1,3 +1,3 @@
+export * from "./Dropdown"
+export * from './Accordion'
 export { BaseButton, ThemedButton } from "./Button";
-export { ClickDropdown, HoverDropdown } from "./Dropdown"
-export { AccordionMenu } from './Accordion'
