+++ styled.d.ts
@@ -1,12 +1,9 @@
 // import original module declarations
 import 'styled-components';
-import { ExpandedThemeInterface, OutputTheme } from './src/themes/interfaces'
+import { OutputTheme } from './src/themes'
 
 // and extend them!
 declare module 'styled-components' {
-  // export interface DefaultTheme extends ExpandedThemeInterface {
-      
-  // }
 
   export interface DefaultTheme{
     presets: {[key:string]: string},
